HeroProfile & HeroProfilePage結構簡化
