--- src/components/HeroProfile/HeroProfile.style.ts
@@ -1,3 +1,0 @@
-import styled from 'styled-components';
-
-export const StyledHeroProfile = styled.div``;
