--- src/components/HeroProfile/HeroProfile.type.ts
@@ -1,6 +1,0 @@
-import { Profile } from "src/types/hero.type";
-
-export interface HeroProfileProps {
-  profile: Profile;
-  onProfileChange?: (profile: Profile) => void;
-}
