+++ src/pages/HeroProfile/HeroProfile.type.ts
@@ -1,0 +1,11 @@
+export interface HeroProfileProps {
+    
+}
+
+export interface Profile {
+	str: number;
+	int: number;
+	agi: number;
+	luk: number;
+	total: number;
+}
