--- src/pages/HeroProfilePage/HeroProfilePage.type.ts
@@ -1,3 +1,0 @@
-export interface HeroProfilePageProps {
-    
-}
