+++ src/router/route-config.tsx
@@ -1,5 +1,5 @@
 import HeroListPage from "src/pages/HeroListPage";
-import HeroProfilePage from "src/pages/HeroProfilePage";
+import HeroProfile from "src/pages/HeroProfile";
 import Home from "src/pages/Home";
 import { RouteConfig } from "./AppRouter.type";
 
@@ -17,7 +17,7 @@ export default [
 		childs: [
 			{ 
 				path: '/heroes/:heroId', 
-				element: HeroProfilePage
+				element: HeroProfile
 			}
 		]
 	},
