+++ src/types/hero.type.ts
@@ -3,11 +3,3 @@ export interface Hero {
 	name: string;
 	image: string;
 }
-
-export interface Profile {
-	str: number;
-	int: number;
-	agi: number;
-	luk: number;
-	total: number;
-}
